refactor(presentation): suspend on the urlpattern polyfill with use() instead of an effect gate
The polyfill import promise is cached at module level so its identity is
stable across render attempts, and only created when a studio configures
mainDocuments resolvers in a runtime without native URLPattern. The
suspension is caught by the studio's tool-level Suspense boundary, the
same one that covers the presentation tool's own lazy chunk, and a
failed polyfill load now surfaces in the tool error boundary instead of
being swallowed by a console.error.
